refactor: 파일명 수정, 메서드명 수정
- app.js -> LottoApp.js 로 수정
- lottoLi -> lottoTicketAndNumberElement 로 수정
