+++ src/index.js
@@ -1,4 +1,4 @@
 import './css/index';
-import LottoApp from './js/app';
+import LottoApp from './js/LootoApp';
 
 new LottoApp('#app');
