+++ src/js/LootoApp.js
